continue to refactor and review react... this is really one of the best reviews I have ever had
