+++ components/movieList.js
@@ -2,10 +2,16 @@ import React from "react";
 
 class MovieList extends React.Component {
 
+  shorten = (text, maxLength) => {
+    if (text && text.length > maxLength) {
+      return text.substr(0, maxLength) + '...'
+    }
+  }
+
   renderMovies(movies) {
     return movies.map((movie) => 
     (
-      <div className="col-lg-4 col-md-6 mb-4">
+      <div className="col-lg-4 col-md-6 mb-4" key={movie.id}>
         <div className="card h-100">
           <a href="#">
             <img className="card-img-top" src={movie.cover} alt="" />
@@ -15,7 +21,7 @@ class MovieList extends React.Component {
               <a href="#">{movie.name}</a>
             </h4>
             <h5>$24.99</h5>
-            <p className="card-text">{movie.description}</p>
+            <p className="card-text">{this.shorten(movie.longDesc, 75)})</p>
           </div>
           <div className="card-footer">
             {parseInt(movie.rating) >= 5 ? (
